requirements.txt
colorama
webbrowser
ctypes
pyautogui
os 
time
pynput.keyboard